# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dokoopy 1.0.0",
    "description": "API for charity landing page of 57th brigade.\n\nThis Server based on the OpenAPI 3.0 specification\n\nSome useful links:\n- [The Dokoopy-backend repository](https://github.com/baza-trainee/dokoopy-backend)",
    "contact": {
      "name": "API developers",
      "email": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/api",
      "description": "Development Server"
    },
    {
      "url": "https://dokoopy.onrender.com/api",
      "description": "Production Server"
    },
    {
      "url": "https://dokoopy-pr-19.onrender.com/api",
      "description": "Test Server for feature get-main-content"
    }
  ],
  "consumes": ["application/json", "multipart/form-data", "text/html"],
  "produces": ["application/json", "multipart/form-data", "text/html"],
...
  "paths": {
...
    "/projects/admin": {
      "post": {
        "summary": "Add new project",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new project",
        "operationId": "addProject",
        "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/AddProjectRequest"
                }
              }
            }
        },
        "responses": {
          "201": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddProjectResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "get": {
        "summary": "Get all projects for the admin panel",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get all project for the admin panel",
        "operationId": "getAllProjects",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllProjectsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/projects/admin/{id}": {
      "patch": {
        "summary": "Update project info",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update project information in the database",
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "delete": {
          "tags": [
            "Projects"
          ],
          "security": [
            {
              "Bearer": []
            }
          ],
          "summary": "Delete the project",
          "description": "Delete the project from the database",
          "operationId": "deleteProject",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of project that needs to be deleted",
              "required": true,
              "schema": {
                "type": "string",
                "description": "Delete the project with the specified ID"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Project deleted"
            },
            "401": {
              "description": "Not authorized"
            }
          }
      }
    },
    "/projects": {
      "get": {
        "summary": "Get all projects for the main page",
        "tags": [
          "Projects"
        ],
        "description": "Get all project for the main page",
        "operationId": "getAllProjects",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllProjectsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/partners/admin": {
      "post": {
        "summary": "Add new partner",
        "tags": [
          "Partners"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new partner",
        "operationId": "addPartner",
        "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/AddPartnerRequest"
                }
              }
            }
        },
        "responses": {
          "201": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddPartnerResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "get": {
        "summary": "Get all partners for the admin panel",
        "tags": [
          "Partners"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get all partners for the admin panel",
        "operationId": "getAllPartners",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllPartnersResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/partners/admin{id}": {
      "patch": {
        "summary": "Update partner info",
        "tags": [
          "Partners"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update partner information in the database",
        "operationId": "updatePartner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartnerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddPartnerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "delete": {
          "tags": [
            "Partners"
          ],
          "security": [
            {
              "Bearer": []
            }
          ],
          "summary": "Delete the partner",
          "description": "Delete the partner from the database",
          "operationId": "deletePartner",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of partner that needs to be deleted",
              "required": true,
              "schema": {
                "type": "string",
                "description": "Delete the partner with the specified ID"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Partner deleted"
            },
            "401": {
              "description": "Not authorized"
            }
          }
      }
    },
    "/partners": {
      "get": {
        "summary": "Get all partners for the main page",
        "tags": [
          "Partners"
        ],
        "description": "Get all partners for the main page",
        "operationId": "getAllPartners",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllPartnersResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
...
    "/reports/admin": {
      "post": {
        "summary": "Add new report",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new report",
        "operationId": "addReport",
        "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/AddReportRequest"
                }
              }
            }
        },
        "responses": {
          "201": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddReportResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      },
      "get": {
        "summary": "Get all reports for the admin panel",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get all reports for the admin panel",
        "operationId": "getAllReports",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllReportsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/reports/admin/{id}": {
      "patch": {
        "summary": "Update report",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update report in the database",
        "operationId": "updateReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      }
    },
    "/reports": {
      "get": {
        "summary": "Get all reports for the main page",
        "tags": [
          "Reports"
        ],
        "description": "Get all reports for the main page",
        "operationId": "getAllReports",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllReportsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddProjectRequest": {
        "type": "object",
        "required": [
          "title",
          "title_eng",
          "description",
          "description_eng",
          "imageURL"
        ],
        "properties": {
          "title": {
              "type": "string",
              "example": "Проєкт 1"
          },
          "title_eng": {
            "type": "string",
            "example": "Project 1"
          },
            "description": {
              "type": "string",
              "example": "Це опис проєкту"
          },
            "description_eng": {
              "type": "string",
              "example": "This is the project description"
          },
            "imageURL": {
              "type": "string",
              "format": "binary"
          }
        }
      },
      "AddProjectResponse": {
          "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "Backend-generated unique identifier",
                "example": "GGHJlhkj668__hjkhjkhh$%"
              },
              "title": {
                "type": "string",
                "description": "Project title in ukrainian",
                "example": "Проєкт 1"
              },
              "title_eng": {
                "type": "string",
                "description": "Project title in english",
                "example": "Project 1"
              },
              "description": {
                "type": "string",
                "description": "Project description",
                "example": "Це опис проекту"
              },
              "description_eng": {
                "type": "string",
                "example": "This is the project description"
              },
              "imageURL": {
                  "type": "string",
                  "description": "Project image url",
                  "example": "https://res.cloudinary.com/dwvdhkbkm/image/upload/v1697058515/dokoopy/o1tvpkyfeygpazxhuqba.jpg"
              },
              "createdAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              }
          }
      },
      "UpdateProjectRequest": {
                "type": "object",
                "properties": {
                    "title": {
                      "type": "string",
                      "example": "Проєкт 1"
                    },
                    "title_eng": {
                      "type": "string",
                      "example": "Project 1"
                    },
                    "description": {
                      "type": "string",
                      "example": "Це перший проєкт"
                    },
                    "description_eng": {
                      "type": "string",
                      "example": "This is the first project"
                    },
                    "imageURL": {
                        "type": "string",
                        "format": "binary"
                    }
                }
      },
      "AddPartnerRequest": {
        "type": "object",
        "required": [
          "title",
          "link",
          "imageURL"
        ],
        "properties": {
          "title": {
              "type": "string",
              "example": "Partner 1"
          },
            "link": {
              "type": "string",
              "example": "https://olx.ua"
          },
            "imageURL": {
              "type": "string",
              "format": "binary"
          }
        }
      },
      "AddPartnerResponse": {
          "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "Backend-generated unique identifier",
                "example": "GGHJlhkj668__hjkhjkhh$%"
              },
              "title": {
                "type": "string",
                "description": "Partner title",
                "example": "Partner 1"
              },
              "link": {
                "type": "string",
                "description": "Partner link",
                "example": "https://olx.ua"
              },
              "imageURL": {
                  "type": "string",
                  "description": "Partner image url",
                  "example": "partners\\qtc7YIPFu7y55n-hrA0Nl_IMG_7842.JPG"
              },
              "createdAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              }
          }
      },
      "UpdatePartnerRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Partner 1"
                    },
                    "link": {
                        "type": "string",
                        "example": "https://olx.ua"
                    },
                    "imageURL": {
                        "type": "string",
                        "format": "binary"
                    }
            }
      },
...
      "AddReportRequest": {
        "type": "object",
        "required": [
          "reportURL"
        ],
        "properties": {
          "reportURL": {
              "type": "string",
              "format": "binary"
          }
        }
      },
      "AddReportResponse": {
          "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "Backend-generated unique identifier",
                "example": "GGHJlhkj668__hjkhjkhh$%"
              },
              "reportURL": {
                "type": "string",
                "description": "Report link",
                "example": "reports\\eYLKBYxgM-MVDo31zUgnz_test.pdf"
              }
          }
      },
      "UpdateReportRequest": {
        "type": "object",
        "required": [
          "reportURL"
        ],
        "properties": {
          "reportURL": {
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "GetAllProjectsResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "description": "Backend-generated unique identifier",
                  "example": "GGHJlhkj668__hjkhjkhh$%"
                },
                "title": {
                  "type": "string",
                  "description": "The title of project",
                  "example": "Проєкт 1"
                },
                "title_eng": {
                  "type": "string",
                  "description": "The title of project",
                  "example": "Project 1"
                },
                "description": {
                  "type": "string",
                  "description": "The description of project",
                  "example": "Це перший проєкт"
                },
                "description_eng": {
                  "type": "string",
                  "description": "The description of project",
                  "example": "This is the first project"
                },
                "imageURL": {
                  "type": "string",
                  "description": "Link on project's photo",
                  "example": "projects\\-SWk2ERNOBwlHLgJVpW07_IMG_7856.JPG"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                }
              }
            },
            "description": "List of projects found by request"
          }
        }
      },
      "GetAllPartnersResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "description": "Backend-generated unique identifier",
                  "example": "GGHJlhkj668__hjkhjkhh$%"
                },
                "title": {
                  "type": "string",
                  "description": "The name of partner",
                  "example": "Partner 1"
                },
                "link": {
                  "type": "string",
                  "description": "The link on partner site",
                  "example": "https://baza-trainee.tech/ua"
                },
                "imageURL": {
                  "type": "string",
                  "description": "Link on project's photo",
                  "example": "partners\\-SWk2ERNOBwlHLgJVpW07_IMG_7856.JPG"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                }
              }
            },
            "description": "List of partners found by request"
          }
        }
      },
      "GetAllReportsResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "description": "Backend-generated unique identifier",
                  "example": "GGHJlhkj668__hjkhjkhh$%"
                },
                "reportURL": {
                  "type": "string",
                  "description": "Link on project's photo",
                  "example": "https://res.cloudinary.com/dwvdhkbkm/image/upload/v1697059048/dokoopy/vlxpr3kfm2fwzjoungpo.pdf"
                }
              }
            },
            "description": "List of reports found by request"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}